break higher-level Invariant into its own file
